Updated Diffusion Examples to 6.11.9-134649.30a0d0be-SNAPSHOT

+++ apple/Features/TimeSeries/TimeSeriesRangeQueryExample.m
@@ -119,7 +119,7 @@
          }
 
          NSString *const author = eventMetadata.author;
-         NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:eventMetadata.timestamp / 1000];
+         NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:(double)eventMetadata.timestamp / 1000.0];
          NSDateFormatter *const dateFormatter = [[NSDateFormatter alloc] init];
          [dateFormatter setDateFormat:@"yyyy-MM-dd hh:mm:ss"];
          NSUInteger sequence = [NSNumber numberWithUnsignedLongLong:eventMetadata.sequence].unsignedIntegerValue;
@@ -153,7 +153,7 @@
          }
 
          NSString *const author = eventMetadata.author;
-         NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:eventMetadata.timestamp / 1000];
+         NSDate *const timestamp = [NSDate dateWithTimeIntervalSince1970:(double)eventMetadata.timestamp / 1000.0];
          NSDateFormatter *const dateFormatter = [[NSDateFormatter alloc] init];
          [dateFormatter setDateFormat:@"yyyy-MM-dd hh:mm:ss"];
          NSUInteger sequence = [NSNumber numberWithUnsignedLongLong:eventMetadata.sequence].unsignedIntegerValue;
